+++ CHANGELOG.md
@@ -12,6 +12,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `Get-SqlDscConfigurationOption` - Returns the available configuration
       options that can be used with the DSC resource _SqlConfiguration_.
 
+### Changed
+
+- SqlServerDsc
+  - Re-enable integration tests for dbatools.
+  - Bumped dbatools to v2.0.1 for the integration tests.
+
+## [16.3.1] - 2023-05-06
+
+### Changed
+
+- SqlServerDsc
+  - Temporary disable integration tests for dbatools.
+- SqlAg
+  - Added optional parameter `SeedingMode` that will set the SeedingMode for the
+    SQL Server 2016 and higher. This parameter can only be used together with the
+    module _SqlServer_ installed (tested  v21.0.17099). The parameter will be
+    ignored if SQLPS module will be used.
+- SqlAgReplica
+  - Added optional parameter `SeedingMode` that will set the SeedingMode for the
+    SQL Server 2016 and higher ([issue #487](https://github.com/dsccommunity/SqlServerDsc/issues/487)).
+    This parameter can only be used together with the module _SqlServer_ installed
+    (tested v21.0.17099). The parameter will be ignored if SQLPS module will be
+    used.
+
+### Fixed
+
+- `Import-SqlDscPreferredModule`
+  - Now the command does not fail when checking if SQLPS is loaded into the
+    session ([issue #1928](https://github.com/dsccommunity/SqlServerDsc/issues/1928)).
+
 ## [16.3.0] - 2023-04-26
 
 ### Remove